+++ packages/config/src/Config.ts
@@ -508,10 +508,11 @@ export class Config {
      *
      * @internal
      * @param maskSecure Indicates whether we should mask off secure properties.
+     * @param excludeGlobalLayer Indicates whether we should exclude global layers.
      *
      * @returns The resulting Config object
      */
-    private layerMerge(maskSecure?: boolean): IConfig {
+    public layerMerge(maskSecure?: boolean, excludeGlobalLayer?: boolean): IConfig {
         // config starting point
         // NOTE: "properties" and "secure" only apply to the individual layers
         // NOTE: they will be blank for the merged config
@@ -548,9 +549,11 @@ export class Config {
 
         // Traverse all the global profiles merging any missing from project profiles
         c.profiles = proj;
-        for (const [n, p] of Object.entries(glbl)) {
-            if (c.profiles[n] == null)
-                c.profiles[n] = p;
+        if (!excludeGlobalLayer) {
+            for (const [n, p] of Object.entries(glbl)) {
+                if (c.profiles[n] == null)
+                    c.profiles[n] = p;
+            }
         }
 
         return c;