+++ packages/config/src/doc/IProfInfoUpdatePropOpts.ts
@@ -10,6 +10,7 @@
 */
 
 import { IProfArgValue } from "./IProfArgAttrs";
+import { IProfLocOsLoc } from "./IProfLoc";
 import { IProfMergedArg } from "./IProfMergedArg";
 
 /**
@@ -35,6 +36,12 @@ export interface IProfInfoUpdateKnownPropOpts extends IProfInfoUpdatePropCommonO
      * Merged arguments object describing the location of the property to update
      */
     mergedArgs: IProfMergedArg;
+
+    /**
+     * Optional osLoc information.
+     * This will allow the updateKnownProperty function to switch active layer if needed
+     */
+    osLocInfo?: IProfLocOsLoc;
 }
 
 export interface IProfInfoUpdatePropCommonOpts {